requirements.txt
python-dotenv
RPi.GPIO